Refactor gateways drawer components
