+++ client/src/pages/Gateways.test.tsx
@@ -105,10 +105,62 @@ describe("Gateways", () => {
     renderWithProviders(<Gateways />);
 
     expect(screen.getByRole("heading", { name: "Virtual servers" })).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Create Server" })).toBeInTheDocument();
     expect(screen.getByText("GH repo tasks")).toBeInTheDocument();
     expect(screen.getByText("6")).toBeInTheDocument();
     expect(screen.getByText("team")).toBeInTheDocument();
     expect(screen.queryByText("MCP server")).not.toBeInTheDocument();
+  });
+
+  it("navigates to the source form when the header create server button is clicked", async () => {
+    const user = userEvent.setup();
+    const mockServer: VirtualServer = {
+      id: "gateway-1",
+      name: "GH repo tasks",
+      description: "Test server",
+      icon: "",
+      createdAt: "2026-04-16T13:23:12Z",
+      updatedAt: "2026-04-16T13:23:12Z",
+      enabled: true,
+      associatedTools: [],
+      associatedToolIds: [],
+      associatedResources: [],
+      associatedPrompts: [],
+      associatedA2aAgents: [],
+      metrics: null,
+      tags: [],
+      createdBy: "admin@example.com",
+      createdFromIp: "127.0.0.1",
+      createdVia: "ui",
+      createdUserAgent: "Mozilla/5.0",
+      modifiedBy: null,
+      modifiedFromIp: null,
+      modifiedVia: null,
+      modifiedUserAgent: null,
+      importBatchId: null,
+      federationSource: null,
+      version: 1,
+      teamId: "team-1",
+      team: "Test Team",
+      ownerEmail: "admin@example.com",
+      visibility: "team",
+      oauthEnabled: false,
+      oauthConfig: null,
+    };
+
+    mockUseQuery.mockReturnValue({
+      data: { servers: [mockServer] },
+      error: null,
+      isLoading: false,
+      execute: vi.fn(),
+      refetch: vi.fn(),
+    });
+
+    renderWithProviders(<Gateways />);
+
+    await user.click(screen.getByRole("button", { name: "Create Server" }));
+
+    expect(mockNavigate).toHaveBeenCalledWith("/app/servers?openForm=true");
   });
 
   it("navigates to servers page with open parameter when MCP server connect is clicked", async () => {
@@ -155,7 +207,7 @@ describe("Gateways", () => {
   it("opens virtual server details and keeps unfinished row actions disabled", async () => {
     const user = userEvent.setup();
     const mockServer: VirtualServer = {
-      id: "gateway-1",
+      id: "gateway/1?mode=detail",
       name: "GH repo tasks",
       description: "Test server",
       icon: "",
@@ -199,7 +251,7 @@ describe("Gateways", () => {
     };
 
     mockUseQuery.mockImplementation((path) => {
-      if (path === "/servers/gateway-1") {
+      if (path === "/servers/gateway%2F1%3Fmode%3Ddetail") {
         return {
           data: detailServer,
           error: null,
@@ -234,7 +286,9 @@ describe("Gateways", () => {
 
     await user.click(viewDetails);
 
-    expect(mockUseQuery).toHaveBeenCalledWith("/servers/gateway-1", { enabled: true });
+    expect(mockUseQuery).toHaveBeenCalledWith("/servers/gateway%2F1%3Fmode%3Ddetail", {
+      enabled: true,
+    });
     expect(screen.getByText("Virtual server details")).toBeInTheDocument();
     expect(
       screen.getByText(
@@ -248,6 +302,12 @@ describe("Gateways", () => {
     expect(screen.getByText("Server ID")).toBeInTheDocument();
     expect(screen.getByLabelText("Copy URL")).toBeInTheDocument();
     expect(screen.getByText("Activity")).toBeInTheDocument();
+    const drawerAddSourcesButton = screen.getByRole("button", { name: "Add sources" });
+    expect(drawerAddSourcesButton).toBeInTheDocument();
+    await user.click(drawerAddSourcesButton);
+    expect(mockNavigate).toHaveBeenCalledWith("/app/servers?openForm=true");
+    await user.click(screen.getByRole("button", { name: "Add components" }));
+    expect(mockNavigate).toHaveBeenCalledWith("/app/servers?openForm=true");
     expect(screen.getByText("Get Repo Issues")).toBeInTheDocument();
     expect(screen.getByText("GITHUB_GET_REPO_ISSUES")).toBeInTheDocument();
     expect(screen.getAllByText("github://repo/{owner}/{repo}").length).toBeGreaterThan(0);
@@ -258,6 +318,13 @@ describe("Gateways", () => {
     expect(screen.getByText("Create New Issue")).toBeInTheDocument();
     expect(screen.queryByText("github://repo/{owner}/{repo}")).not.toBeInTheDocument();
     expect(screen.queryByText("summarize_pull_request")).not.toBeInTheDocument();
+
+    await user.click(screen.getByRole("button", { name: "Close virtual server details" }));
+    await user.click(screen.getByRole("button", { name: "Actions for GH repo tasks" }));
+    await user.click(await screen.findByRole("menuitem", { name: "View details" }));
+
+    expect(screen.getAllByText("github://repo/{owner}/{repo}").length).toBeGreaterThan(0);
+    expect(screen.getAllByText("summarize_pull_request").length).toBeGreaterThan(0);
   });
 
   it("renders virtual server card without crashing when array fields are missing", () => {
